% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajust_work_variables.R
\name{ajust_work_variables}
\alias{ajust_work_variables}
\title{Implements a correction for the variable "class worker"}
\usage{
ajust_work_variables(CensusData)
}
\description{
Implements a correction for the variable "class worker"
}
