% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_demographics_age_2000.R
\name{build_demographics_age_2000}
\alias{build_demographics_age_2000}
\title{Builds a synthetic variable for age - 2000}
\usage{
build_demographics_age_2000(CensusData)
}
\description{
Builds a synthetic variable for age - 2000
}
