% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_demographics_householdStatus_1960.R
\name{build_demographics_householdStatus_1960}
\alias{build_demographics_householdStatus_1960}
\title{Builds household dweller status for 1960}
\usage{
build_demographics_householdStatus_1960(CensusData)
}
\description{
Builds household dweller status for 1960
}
