% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_demographics_householdStatus_1970.R
\name{build_demographics_householdStatus_1970}
\alias{build_demographics_householdStatus_1970}
\title{Builds household dweller status for 1970}
\usage{
build_demographics_householdStatus_1970(CensusData)
}
\description{
Builds household dweller status for 1970
}
