% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_demographics_householdStatus_1980.R
\name{build_demographics_householdStatus_1980}
\alias{build_demographics_householdStatus_1980}
\title{Builds household dweller status for 1980}
\usage{
build_demographics_householdStatus_1980(CensusData)
}
\description{
Builds household dweller status for 1980
}
