% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_demographics_householdStatus_1991.R
\name{build_demographics_householdStatus_1991}
\alias{build_demographics_householdStatus_1991}
\title{Builds household dweller status for 1991}
\usage{
build_demographics_householdStatus_1991(CensusData)
}
\description{
Builds household dweller status for 1991
}
