% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_demographics_householdStatus_2000.R
\name{build_demographics_householdStatus_2000}
\alias{build_demographics_householdStatus_2000}
\title{Builds household dweller status for 2000}
\usage{
build_demographics_householdStatus_2000(CensusData)
}
\description{
Builds household dweller status for 2000
}
