% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_demographics_householdStatus_2010.R
\name{build_demographics_householdStatus_2010}
\alias{build_demographics_householdStatus_2010}
\title{Builds household dweller status for 2010}
\usage{
build_demographics_householdStatus_2010(CensusData)
}
\description{
Builds household dweller status for 2010
}
