% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_demographics_male_1970.R
\name{build_demographics_male_1970}
\alias{build_demographics_male_1970}
\title{Builds a synthetic variable for age - 1960}
\usage{
build_demographics_male_1970(CensusData)
}
\description{
Builds a synthetic variable for age - 1960
}
