% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_demographics_nonrelative_2010.R
\name{build_demographics_nonrelative_2010}
\alias{build_demographics_nonrelative_2010}
\title{Builds a synthetic variable for age - 1960}
\usage{
build_demographics_nonrelative_2010(CensusData)
}
\description{
Builds a synthetic variable for age - 1960
}
