% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_demographics_residenceStatus_1980.R
\name{build_demographics_residenceStatus_1980}
\alias{build_demographics_residenceStatus_1980}
\title{Builds a synthetic variable for age - 1960}
\usage{
build_demographics_residenceStatus_1980(CensusData)
}
\description{
Builds a synthetic variable for age - 1960
}
