% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_education_educationAttainment_1960.R
\name{build_education_educationAttainment_1960}
\alias{build_education_educationAttainment_1960}
\title{Builds a synthetic variable for educationAttainment attainment - 1960}
\usage{
build_education_educationAttainment_1960(CensusData)
}
\description{
Builds a synthetic variable for educationAttainment attainment - 1960
}
