% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_education_educationAttainment_1991.R
\name{build_education_educationAttainment_1991}
\alias{build_education_educationAttainment_1991}
\title{Builds a synthetic variable for educationAttainment attainment - 1991}
\usage{
build_education_educationAttainment_1991(CensusData)
}
\arguments{
\item{data.frame}{}
}
\description{
Builds a synthetic variable for educationAttainment attainment - 1991
}
