% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_education_levelattnd_1970.R
\name{build_education_levelattnd_1970}
\alias{build_education_levelattnd_1970}
\title{Builds a synthetic variable for education attainment - 1970}
\usage{
build_education_levelattnd_1970(CensusData)
}
\description{
Builds a synthetic variable for education attainment - 1970
}
