% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_education_literacy_2000.R
\name{build_education_literacy_2000}
\alias{build_education_literacy_2000}
\title{Builds a synthetic variable for education attainment - 2000}
\usage{
build_education_literacy_2000(CensusData)
}
\description{
Builds a synthetic variable for education attainment - 2000
}
