% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_education_yearsOfSchooling_1970.r
\name{build_education_yearsOfSchooling_1970}
\alias{build_education_yearsOfSchooling_1970}
\title{Builds a synthetic variable for educationAttainment attainment - 1970}
\usage{
build_education_yearsOfSchooling_1970(CensusData)
}
\description{
Builds a synthetic variable for educationAttainment attainment - 1970
}
