% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_education_yearsOfSchooling_1980.R
\name{build_education_yearsOfSchooling_1980}
\alias{build_education_yearsOfSchooling_1980}
\title{Builds a synthetic variable for yearsOfSchooling attainment - 1980}
\usage{
build_education_yearsOfSchooling_1980(CensusData)
}
\description{
Builds a synthetic variable for yearsOfSchooling attainment - 1980
}
