% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_education_yearsOfSchooling_2010.R
\name{build_education_yearsOfSchooling_2010}
\alias{build_education_yearsOfSchooling_2010}
\title{Builds a synthetic variable for yearsOfSchooling attainment - 2010}
\usage{
build_education_yearsOfSchooling_2010(CensusData)
}
\description{
Builds a synthetic variable for yearsOfSchooling attainment - 2010
}
