% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_geography_minCompAreas1970to2010_1991.R
\name{build_geography_minCompAreas1970to2010_1991}
\alias{build_geography_minCompAreas1970to2010_1991}
\title{Builds a synthetic variable for education attainment - 2010}
\usage{
build_geography_minCompAreas1970to2010_1991(CensusData)
}
\description{
Builds a synthetic variable for education attainment - 2010
}
