% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_geography_municipalityCurrent_1960.R
\name{build_geography_municipalityCurrent_1960}
\alias{build_geography_municipalityCurrent_1960}
\title{Harmonize a set of education variables all at once}
\usage{
build_geography_municipalityCurrent_1960(CensusData)
}
\description{
Harmonize a set of education variables all at once
}
