% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_geography_municipalityCurrent_1970.R
\name{build_geography_municipalityCurrent_1970}
\alias{build_geography_municipalityCurrent_1970}
\title{Harmonize a set of education variables all at once}
\usage{
build_geography_municipalityCurrent_1970(CensusData,
  state_var_name = "uf")
}
\description{
Harmonize a set of education variables all at once
}
