% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_geography_municipalityCurrent_1980.R
\name{build_geography_municipalityCurrent_1980}
\alias{build_geography_municipalityCurrent_1980}
\title{Harmonize a set of education variables all at once}
\usage{
build_geography_municipalityCurrent_1980(CensusData)
}
\description{
Harmonize a set of education variables all at once
}
