% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_geography_regionMinimumComparable.R
\name{build_geography_regionMinimumComparable}
\alias{build_geography_regionMinimumComparable}
\title{Builds a synthetic variable for education attainment - 2010}
\usage{
build_geography_regionMinimumComparable(CensusData)
}
\description{
Builds a synthetic variable for education attainment - 2010
}
