% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_geography_regionMiniumComparable.R
\name{build_geography_regionMiniumComparable}
\alias{build_geography_regionMiniumComparable}
\title{Builds a synthetic variable for education attainment - 2010}
\usage{
build_geography_regionMiniumComparable(CensusData)
}
\description{
Builds a synthetic variable for education attainment - 2010
}
