% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_geography_state2010standard_2000.R
\name{build_geography_state2010standard_2000}
\alias{build_geography_state2010standard_2000}
\title{Builds a synthetic variable for education attainment - 2010}
\usage{
build_geography_state2010standard_2000(CensusData)
}
\description{
Builds a synthetic variable for education attainment - 2010
}
