% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_geography_stateMinimumComparable_1960.R
\name{build_geography_stateMinimumComparable_1960}
\alias{build_geography_stateMinimumComparable_1960}
\title{Builds a synthetic variable for education attainment - 2010}
\usage{
build_geography_stateMinimumComparable_1960(CensusData)
}
\description{
Builds a synthetic variable for education attainment - 2010
}
