% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_household_electricity.R
\name{build_household_electricity}
\alias{build_household_electricity}
\title{Builds a synthetic variable for age - 1970}
\usage{
build_household_electricity(CensusData)
}
\description{
Builds a synthetic variable for age - 1970
}
