% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_household_garbageCollection.R
\name{build_household_garbageCollection}
\alias{build_household_garbageCollection}
\title{Builds a synthetic variable for age - 1970}
\usage{
build_household_garbageCollection(CensusData)
}
\description{
Builds a synthetic variable for age - 1970
}
