% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_household_hhType.R
\name{build_household_hhType}
\alias{build_household_hhType}
\title{Builds a synthetic variable for age - 1970}
\usage{
build_household_hhType(CensusData)
}
\description{
Builds a synthetic variable for age - 1970
}
