% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_household_numberDwellers_1970.R
\name{build_household_numberDwellers_1970}
\alias{build_household_numberDwellers_1970}
\title{Builds a synthetic variable for education attainment - 2010}
\usage{
build_household_numberDwellers_1970(CensusData)
}
\description{
Builds a synthetic variable for education attainment - 2010
}
