% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_household_properSewage.R
\name{build_household_properSewage}
\alias{build_household_properSewage}
\title{Builds a synthetic variable for age - 1970}
\usage{
build_household_properSewage(CensusData)
}
\description{
Builds a synthetic variable for age - 1970
}
