% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_household_publicWater.R
\name{build_household_publicWater}
\alias{build_household_publicWater}
\title{Builds a synthetic variable for age - 1970}
\usage{
build_household_publicWater(CensusData)
}
\description{
Builds a synthetic variable for age - 1970
}
