% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_household_househouldParticular_1970.R
\name{build_househould_househouldParticular_1970}
\alias{build_househould_househouldParticular_1970}
\title{Builds a synthetic variable for age - 1970}
\usage{
build_househould_househouldParticular_1970(CensusData)
}
\description{
Builds a synthetic variable for age - 1970
}
