% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_identification_year_1960.R
\name{build_identification_year_1960}
\alias{build_identification_year_1960}
\title{Builds a synthetic variable for year - 1960}
\usage{
build_identification_year_1960(CensusData)
}
\description{
Builds a synthetic variable for year - 1960
}
