% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_identification_year_1970.R
\name{build_identification_year_1970}
\alias{build_identification_year_1970}
\title{Builds a synthetic variable for year - 1970}
\usage{
build_identification_year_1970(CensusData)
}
\description{
Builds a synthetic variable for year - 1970
}
