% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_identification_year_1980.R
\name{build_identification_year_1980}
\alias{build_identification_year_1980}
\title{Builds a synthetic variable for year - 1980}
\usage{
build_identification_year_1980(CensusData)
}
\description{
Builds a synthetic variable for year - 1980
}
