% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_identification_year_1991.R
\name{build_identification_year_1991}
\alias{build_identification_year_1991}
\title{Builds a synthetic variable for year - 1991}
\usage{
build_identification_year_1991(CensusData)
}
\description{
Builds a synthetic variable for year - 1991
}
