% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_identification_year_2000.R
\name{build_identification_year_2000}
\alias{build_identification_year_2000}
\title{Builds a synthetic variable for year - 2000}
\usage{
build_identification_year_2000(CensusData)
}
\description{
Builds a synthetic variable for year - 2000
}
