% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_identification_year_2010.R
\name{build_identification_year_2010}
\alias{build_identification_year_2010}
\title{Builds a synthetic variable for year - 2010}
\usage{
build_identification_year_2010(CensusData)
}
\description{
Builds a synthetic variable for year - 2010
}
