% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_income_hhIncome2010Values.R
\name{build_income_hhIncome2010Values}
\alias{build_income_hhIncome2010Values}
\title{Builds a synthetic variable for education attainment - 2010}
\usage{
build_income_hhIncome2010Values(CensusData)
}
\description{
Builds a synthetic variable for education attainment - 2010
}
