% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_income_indIncomeNominal_1970.R
\name{build_income_indIncomeNominal_1970}
\alias{build_income_indIncomeNominal_1970}
\title{Harmonize a set of income variables all at once}
\usage{
build_income_indIncomeNominal_1970(CensusData)
}
\description{
Harmonize a set of income variables all at once
}
