% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_income_labourIncome2010Values.R
\name{build_income_labourIncome2010Values}
\alias{build_income_labourIncome2010Values}
\title{Harmonize a set of income variables all at once}
\usage{
build_income_labourIncome2010Values(CensusData)
}
\description{
Harmonize a set of income variables all at once
}
