% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_income_totalIncome1960brackets.R
\name{build_income_totalIncome1960brackets}
\alias{build_income_totalIncome1960brackets}
\title{Harmonize a set of income variables all at once}
\usage{
build_income_totalIncome1960brackets(CensusData)
}
\description{
Harmonize a set of income variables all at once
}
