% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_migration_bornSameMunicipality_1991.R
\name{build_migration_bornSameMunicipality_1991}
\alias{build_migration_bornSameMunicipality_1991}
\title{Builds a synthetic variable for education attainment - 2010}
\usage{
build_migration_bornSameMunicipality_1991(CensusData)
}
\description{
Builds a synthetic variable for education attainment - 2010
}
