% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_migration_stateOfBirthMCA_1970.R
\name{build_migration_stateOfBirthMCA_1970}
\alias{build_migration_stateOfBirthMCA_1970}
\title{Harmonize a set of education variables all at once}
\usage{
build_migration_stateOfBirthMCA_1970(CensusData)
}
\description{
Harmonize a set of education variables all at once
}
