% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_migration_stateOfBirthMCA_2000.R
\name{build_migration_stateOfBirthMCA_2000}
\alias{build_migration_stateOfBirthMCA_2000}
\title{Harmonize a set of education variables all at once}
\usage{
build_migration_stateOfBirthMCA_2000(CensusData)
}
\description{
Harmonize a set of education variables all at once
}
