% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_migration_stateOfBirthMCA_2010.R
\name{build_migration_stateOfBirthMCA_2010}
\alias{build_migration_stateOfBirthMCA_2010}
\title{Harmonize a set of education variables all at once}
\usage{
build_migration_stateOfBirthMCA_2010(CensusData)
}
\description{
Harmonize a set of education variables all at once
}
