% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_work_classWorker.R
\name{build_work_classWorker}
\alias{build_work_classWorker}
\title{Creates a synthetic variable for position in occupation}
\usage{
build_work_classWorker(CensusData)
}
\description{
Creates a synthetic variable for position in occupation
}
