% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_work_occupationalStatus_2000.R
\name{build_work_occupationalStatus_2000}
\alias{build_work_occupationalStatus_2000}
\title{Builds a synthetic variable for age - 1960}
\usage{
build_work_occupationalStatus_2000(CensusData)
}
\description{
Builds a synthetic variable for age - 1960
}
