% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_classWorker_econActivity.R
\name{correct_classWorker_econActivity}
\alias{correct_classWorker_econActivity}
\title{Implements a correction for the variable "class worker"}
\usage{
correct_classWorker_econActivity(CensusData, year, type = "census")
}
\description{
Implements a correction for the variable "class worker"
}
