% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_householdData_1970.R
\name{create_householdData_1970}
\alias{create_householdData_1970}
\title{Builds a synthetic variable for age - 1970}
\usage{
create_householdData_1970(CensusData, state_var_name = "uf")
}
\description{
Builds a synthetic variable for age - 1970
}
