% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_at_least_15hours_work.R
\name{dummy_at_least_15hours_work}
\alias{dummy_at_least_15hours_work}
\title{Identifies people who worked at least 15 hours per week}
\usage{
dummy_at_least_15hours_work(CensusData, year, type = "census")
}
\description{
Identifies people who worked at least 15 hours per week
}
