% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_classWorker.R
\name{harmonize_classWorker}
\alias{harmonize_classWorker}
\title{Creates a synthetic variable for position in occupation}
\usage{
harmonize_classWorker(var_posOcup, var_setor, year, type = "censo")
}
\description{
Creates a synthetic variable for position in occupation
}
