% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_demographics.R
\name{harmonize_demographics}
\alias{harmonize_demographics}
\title{Harmonize a set of demographic variables all at once}
\usage{
harmonize_demographics(CensusData, year, harmonize_identification_first = F,
  scripts_folder = "E:/Dropbox-Ro/Dropbox/Rogerio/Artigos/1 - Em faccao/000 - Padroniza??o de Ocupa??es e Setores/Rogerio/Harmonization Scripts - INCOMPLETE",
  delete_originals = T)
}
\description{
Harmonize a set of demographic variables all at once
}
