% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_economicActivity.R
\name{harmonize_economicActivity}
\alias{harmonize_economicActivity}
\title{Harmonize a set of economic activity variables all at once}
\usage{
harmonize_economicActivity(CensusData, year, type = "census",
  correctForHarmonizedAge = T, delete_originals = T)
}
\description{
Harmonize a set of economic activity variables all at once
}
