% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_education.R
\name{harmonize_education}
\alias{harmonize_education}
\title{Harmonize a set of education variables all at once}
\usage{
harmonize_education(CensusData, year, delete_originals = T)
}
\description{
Harmonize a set of education variables all at once
}
