% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_geography.R
\name{harmonize_geography}
\alias{harmonize_geography}
\title{Harmonize a set of geographic variables all at once}
\usage{
harmonize_geography(CensusData, year, delete_originals = T)
}
\description{
Harmonize a set of geographic variables all at once
}
