% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_identification.R
\name{harmonize_identification}
\alias{harmonize_identification}
\title{Harmonize a set of identification variables all at once}
\usage{
harmonize_identification(CensusData, year, delete_originals = T)
}
\description{
Harmonize a set of identification variables all at once
}
