% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_income.R
\name{harmonize_income}
\alias{harmonize_income}
\title{Harmonize a set of income variables all at once}
\usage{
harmonize_income(CensusData, year,
  scripts_folder = "E:/Dropbox-Ro/Dropbox/Rogerio/Artigos/1 - Em faccao/000 - Padroniza??o de Ocupa??es e Setores/Rogerio/Harmonization Scripts - INCOMPLETE",
  harmonize_demographics_first = F, harmonize_identification_first = F)
}
\description{
Harmonize a set of income variables all at once
}
