% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_to_harmonize.R
\name{prepare_to_harmonize}
\alias{prepare_to_harmonize}
\title{Builds a synthetic variable for age - 1960}
\usage{
prepare_to_harmonize(
  Data,
  type,
  year,
  pnadc_freq = NULL,
  quarter = NULL,
  state_var_name = NULL
)
}
\description{
Builds a synthetic variable for age - 1960
}
