% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setoresToISIC.R
\name{setoresToISIC}
\alias{setoresToISIC}
\title{Converts IBGE sector codes to the ISIC 3.1 standard - ILO-UN}
\usage{
setoresToISIC(setores, tipo, ano = NULL, endereco_matriz = "")
}
\description{
Converts IBGE sector codes to the ISIC 3.1 standard - ILO-UN
}
